Clippy: allow or fix `clippy::unnecessary_cast` lint in src/
In a lot of places, the lint is triggered when casting a constant from the
libgit2 bindings that is created (in libgit2-sys) via `git_enum!`. That macro
will use i32 when `#[cfg(target_env = "msvc")]`, and u32 otherwise, meaning
that a number of the casts that clippy flags as "unnecessary" are indeed
necessary. In those cases, use `#[allow]` with a reason.

Similarly, clippy also lints against casting of `c_int` and `c_uint` to `i32`
and `u32` - while those are usually the same types, the libc documentation
makes it clear that they are not always the same. In those cases too, use
`#[allow]` with a reason.

The only libc-provided type with a consistent rust type is that `size_t` is
always the same as `usize` - fix the places that clippy identified where that
conversion is unneeded. Other fixes include casts when the underlying type is
explicitly set in the libgit2 bindings to be a specific rust type, or when
converting mutable pointers to the same type they already are.

After these changes, running the command
`cargo clippy -- -A clippy::all -W clippy::unnecessary_cast` no longer reports
any hits.

+++ src/buf.rs
@@ -47,7 +47,7 @@ impl Deref for Buf {
         if self.raw.ptr.is_null() {
             return &[];
         }
-        unsafe { slice::from_raw_parts(self.raw.ptr as *const u8, self.raw.size as usize) }
+        unsafe { slice::from_raw_parts(self.raw.ptr as *const u8, self.raw.size) }
     }
 }
 
@@ -56,7 +56,7 @@ impl DerefMut for Buf {
         if self.raw.ptr.is_null() {
             return &mut [];
         }
-        unsafe { slice::from_raw_parts_mut(self.raw.ptr as *mut u8, self.raw.size as usize) }
+        unsafe { slice::from_raw_parts_mut(self.raw.ptr as *mut u8, self.raw.size) }
     }
 }
