+++ src/diff.rs
@@ -591,26 +591,30 @@ impl<'a> DiffFile<'a> {
 
     /// Returns the size of this entry, in bytes
     pub fn size(&self) -> u64 {
-        unsafe { (*self.raw).size as u64 }
+        unsafe { (*self.raw).size }
     }
 
     /// Returns `true` if file(s) are treated as binary data.
     pub fn is_binary(&self) -> bool {
+        #[allow(clippy::unnecessary_cast, reason = "u32 unless compiling for msvc target env")]
         unsafe { (*self.raw).flags & raw::GIT_DIFF_FLAG_BINARY as u32 != 0 }
     }
 
     /// Returns `true` if file(s) are treated as text data.
     pub fn is_not_binary(&self) -> bool {
+        #[allow(clippy::unnecessary_cast, reason = "u32 unless compiling for msvc target env")]
         unsafe { (*self.raw).flags & raw::GIT_DIFF_FLAG_NOT_BINARY as u32 != 0 }
     }
 
     /// Returns `true` if `id` value is known correct.
     pub fn is_valid_id(&self) -> bool {
+        #[allow(clippy::unnecessary_cast, reason = "u32 unless compiling for msvc target env")]
         unsafe { (*self.raw).flags & raw::GIT_DIFF_FLAG_VALID_ID as u32 != 0 }
     }
 
     /// Returns `true` if file exists at this side of the delta.
     pub fn exists(&self) -> bool {
+        #[allow(clippy::unnecessary_cast, reason = "u32 unless compiling for msvc target env")]
         unsafe { (*self.raw).flags & raw::GIT_DIFF_FLAG_EXISTS as u32 != 0 }
     }
 
@@ -680,6 +684,7 @@ impl DiffOptions {
     }
 
     fn flag(&mut self, opt: raw::git_diff_option_t, val: bool) -> &mut DiffOptions {
+        #[allow(clippy::unnecessary_cast, reason = "u32 unless compiling for msvc target env")]
         let opt = opt as u32;
         if val {
             self.raw.flags |= opt;
@@ -1041,7 +1046,7 @@ impl<'a> DiffLine<'a> {
 
     /// Offset in the original file to the content
     pub fn content_offset(&self) -> i64 {
-        unsafe { (*self.raw).content_offset as i64 }
+        unsafe { (*self.raw).content_offset }
     }
 
     /// Content of this line as bytes.
@@ -1049,7 +1054,7 @@ impl<'a> DiffLine<'a> {
         unsafe {
             slice::from_raw_parts(
                 (*self.raw).content as *const u8,
-                (*self.raw).content_len as usize,
+                (*self.raw).content_len,
             )
         }
     }
@@ -1143,7 +1148,7 @@ impl<'a> DiffHunk<'a> {
         unsafe {
             slice::from_raw_parts(
                 (*self.raw).header.as_ptr() as *const u8,
-                (*self.raw).header_len as usize,
+                (*self.raw).header_len,
             )
         }
     }
@@ -1276,13 +1281,13 @@ impl<'a> DiffBinaryFile<'a> {
     /// The binary data, deflated
     pub fn data(&self) -> &[u8] {
         unsafe {
-            slice::from_raw_parts((*self.raw).data as *const u8, (*self.raw).datalen as usize)
+            slice::from_raw_parts((*self.raw).data as *const u8, (*self.raw).datalen)
         }
     }
 
     /// The length of the binary data after inflation
     pub fn inflated_len(&self) -> usize {
-        unsafe { (*self.raw).inflatedlen as usize }
+        unsafe { (*self.raw).inflatedlen }
     }
 }
