+++ src/packbuilder.rs
@@ -256,7 +256,7 @@ impl Binding for PackBuilderStage {
 
 extern "C" fn foreach_c(buf: *const c_void, size: size_t, data: *mut c_void) -> c_int {
     unsafe {
-        let buf = slice::from_raw_parts(buf as *const u8, size as usize);
+        let buf = slice::from_raw_parts(buf as *const u8, size);
 
         let r = panic::wrap(|| {
             let data = data as *mut &mut ForEachCb<'_>;
